+++ ocf_data_sampler/torch_datasets/pvnet_uk_regional.py
@@ -10,8 +10,8 @@ from ocf_data_sampler.config import Configuration, load_yaml_configuration
 from ocf_data_sampler.load.load_dataset import get_dataset_dict
 from ocf_data_sampler.select import fill_time_periods, Location, slice_datasets_by_space, slice_datasets_by_time
 from ocf_data_sampler.utils import minutes
-from ocf_data_sampler.torch_datasets.process_and_combine import process_and_combine_datasets, compute
-from ocf_data_sampler.torch_datasets.valid_time_periods import find_valid_time_periods
+from ocf_data_sampler.join.process_and_combine import process_and_combine_datasets, compute
+from ocf_data_sampler.join.valid_time_periods import find_valid_time_periods
 
 xr.set_options(keep_attrs=True)
