+++ ocf_data_sampler/torch_datasets/site.py
@@ -15,8 +15,8 @@ from ocf_data_sampler.select import (
     slice_datasets_by_time, slice_datasets_by_space
 )
 from ocf_data_sampler.utils import minutes
-from ocf_data_sampler.torch_datasets.process_and_combine import process_and_combine_datasets, compute
-from ocf_data_sampler.torch_datasets.valid_time_periods import find_valid_time_periods
+from ocf_data_sampler.join.process_and_combine import process_and_combine_datasets, compute
+from ocf_data_sampler.join.valid_time_periods import find_valid_time_periods
 
 xr.set_options(keep_attrs=True)
